Refatoração do método adicionarItemSaida
Foi feita a remoção do método da classe SaidaController e a adição do
método na nova classe ItemSaidaController.
