+++ almoxarifado-v0.2.2/aplicacao/clinicas/src/main/java/ufc/npi/clinicas/controller/SaidaController.java
@@ -252,48 +252,6 @@ public class SaidaController {
 		}
 		return new Response().withFailStatus().withErrorMessage(Constants.SAIDA_INCLUIR_MATERIAIS_REMOVER_ERRO);
 	}
-
-	@PostMapping("/api/itemSaida/adicionar")
-	public Response adicionarItemSaida(Integer quantidade, Integer idSaida, Integer idMaterial, Integer idEstoqueLote, Integer idSetor){
-		
-		EstoqueLote estoqueLote = estoqueLoteService.buscarPorId(idEstoqueLote);
-		
-		Material material = estoqueLote.getMaterial();
-		
-		SaidaMaterial saidaMaterial = this.saidaMaterialService.buscarPorId(idSaida);
-		
-		ItemSaida itemSaida = new ItemSaida();
-		itemSaida.setSaidaMaterial(saidaMaterial);
-		itemSaida.setMaterial(material);
-		itemSaida.setQuantidade(quantidade);
-		itemSaida.setLote(estoqueLote.getLote());
-		
-		Setor setor = setorService.buscarPorId(idSetor);
-		EstoqueSetor estoqueSetor = estoqueSetorService.buscarPorSetorEMaterial(setor, material);
-
-		
-		if(this.itemSaidaService.existePorSaidaMaterialEMaterialELote(saidaMaterial, material, estoqueLote.getLote())){
-			
-			return new Response().
-					withFailStatus().
-					withErrorMessage(Constants.SAIDA_INCLUIR_MATERIAIS_ERRO_ITEM_SAIDA_EXISTENTE);
-				
-		}else{
-			if (estoqueSetor == null || itemSaida.getQuantidade() > estoqueSetor.getQuantidade()){
-				return new Response().withFailStatus().withErrorMessage(Constants.SAIDA_INCLUIR_MATERIAIS_ERRO_QUANTIDADE_MAIOR_ESTOQUE);
-			}else if(itemSaida.getQuantidade() > estoqueLote.getQuantidade()){
-				return new Response().withFailStatus().withErrorMessage(Constants.SAIDA_INCLUIR_MATERIAIS_ERRO_QUANTIDADE_MAIOR_ESTOQUE_LOTE);
-			}else if(itemSaida.getQuantidade() <= 0){	
-				return new Response().withFailStatus().withErrorMessage(Constants.SAIDA_INCLUIR_MATERIAIS_ERRO_QUANTIDADE_INVALIDA);
-				
-			}else{
-				itemSaida.setSaidaMaterial(saidaMaterial);
-				itemSaida.setMaterial(material);
-				return new Response().withDoneStatus().withObject(this.itemSaidaService.adicionar(itemSaida))
-						.withSuccessMessage(Constants.SAIDA_INCLUIR_MATERIAIS_SUCESSO_ADICIONADO);
-			}
-		}
-	}
 	
 	@PostMapping("/api/itemSaida/listar")
 	public Response listarItensSaidaPorSaida(SaidaMaterial saidaMaterial){
